+++ src/query/models/types.go
@@ -85,18 +85,31 @@ const (
 type TagOptions interface {
 	// Validate validates these tag options.
 	Validate() error
+
 	// SetMetricName sets the name for the `metric name` tag.
-	SetMetricName(metricName []byte) TagOptions
+	SetMetricName(value []byte) TagOptions
+
 	// MetricName gets the name for the `metric name` tag.
 	MetricName() []byte
+
 	// SetBucketName sets the name for the `bucket label` tag.
-	SetBucketName(metricName []byte) TagOptions
+	SetBucketName(value []byte) TagOptions
+
 	// BucketName gets the name for the `bucket label` tag.
 	BucketName() []byte
+
 	// SetIDSchemeType sets the ID generation scheme type.
-	SetIDSchemeType(scheme IDSchemeType) TagOptions
+	SetIDSchemeType(value IDSchemeType) TagOptions
+
 	// IDSchemeType gets the ID generation scheme type.
 	IDSchemeType() IDSchemeType
+
+	// SetFilters sets tag filters.
+	SetFilters(value Filters) TagOptions
+
+	// Filters gets the tag filters.
+	Filters() Filters
+
 	// Equals determines if two tag options are equivalent.
 	Equals(other TagOptions) bool
 }
@@ -150,3 +163,14 @@ type Metric struct {
 
 // Metrics is a list of individual metrics.
 type Metrics []Metric
+
+// Filters is a set of tag filters.
+type Filters []Filter
+
+// Filter is a regex tag filter.
+type Filter struct {
+	// Name is the name of the series.
+	Name []byte
+	// Filters are a set of regex expressions that filter matched tag values.
+	Filters []*regexp.Regexp
+}
